Address TradeBox component test review feedback
