+++ src/domain/testUtils/emptySyntheticsEventsState.ts
@@ -1,0 +1,44 @@
+import type { SyntheticsEventsContextType } from "context/SyntheticsEvents/types";
+
+const noop = () => undefined;
+
+export const emptySyntheticsEventsState: SyntheticsEventsContextType = {
+  orderStatuses: {},
+  depositStatuses: {},
+  withdrawalStatuses: {},
+  shiftStatuses: {},
+  approvalStatuses: {},
+  pendingOrdersUpdates: {},
+  pendingPositionsUpdates: {},
+  positionIncreaseEvents: [],
+  positionDecreaseEvents: [],
+  pendingExpressTxns: {},
+  gelatoTaskStatuses: {},
+  setPendingExpressTxn: noop,
+  updatePendingExpressTxn: noop,
+  setPendingOrder: noop,
+  setPendingOrderUpdate: noop,
+  setPendingFundingFeeSettlement: noop,
+  setPendingPosition: noop,
+  setPendingDeposit: noop,
+  setPendingWithdrawal: noop,
+  setPendingShift: noop,
+  setOrderStatusViewed: noop,
+  setDepositStatusViewed: noop,
+  setWithdrawalStatusViewed: noop,
+  setShiftStatusViewed: noop,
+  setMultichainTransferProgress: noop,
+
+  multichainSourceChainApprovalStatuses: {},
+  setMultichainSourceChainApprovalsActiveListener: noop,
+  removeMultichainSourceChainApprovalsActiveListener: noop,
+  pendingMultichainFunding: [],
+  setMultichainSubmittedDeposit: noop,
+  setMultichainSubmittedWithdrawal: noop,
+  setMultichainWithdrawalSentTxnHash: noop,
+  setMultichainWithdrawalSentError: noop,
+  updatePendingMultichainFunding: noop,
+  multichainFundingPendingIds: {},
+  removeMultichainFundingPendingIds: noop,
+  setMultichainFundingPendingId: noop,
+};
