+++ src/domain/testUtils/mockPositionInfo.ts
@@ -1,8 +1,11 @@
 import type { PositionInfo } from "domain/synthetics/positions/types";
 import { expandDecimals } from "lib/numbers";
+import { hashedPositionKey } from "sdk/configs/dataStore";
 import type { MarketInfo } from "sdk/utils/markets/types";
 import { getPositionKey } from "sdk/utils/positions";
+import { getMarkPrice } from "sdk/utils/prices/utils";
 import type { TokenData } from "sdk/utils/tokens/types";
+import { convertToTokenAmount } from "sdk/utils/tokens/utils";
 
 import { createMockMarketInfo, createMockMarketsData } from "./mockMarketInfo";
 import { USDC_TOKEN } from "./mockTokens";
@@ -29,24 +32,31 @@ export function createMockPositionInfo({
   isLong = true,
   sizeInUsd = expandDecimals(2000, 30),
   sizeInTokens = expandDecimals(1, 18),
-  collateralAmount = expandDecimals(1000, USDC_TOKEN.decimals),
+  collateralAmount,
   collateralUsd = expandDecimals(1000, 30),
 }: MockPositionInfoOverrides): PositionInfo {
+  const resolvedCollateralAmount =
+    collateralAmount ?? convertToTokenAmount(collateralUsd, collateralToken.decimals, collateralToken.prices.minPrice);
+
+  if (resolvedCollateralAmount === undefined) {
+    throw new Error("Unable to derive mock position collateral amount");
+  }
+
   const key = getPositionKey(account, marketInfo.marketTokenAddress, collateralToken.address, isLong);
-  const markPrice = marketInfo.indexToken.prices.minPrice;
+  const markPrice = getMarkPrice({ prices: marketInfo.indexToken.prices, isIncrease: false, isLong });
   const entryPrice =
     sizeInTokens > 0n ? (sizeInUsd * 10n ** BigInt(marketInfo.indexToken.decimals)) / sizeInTokens : 0n;
   const leverage = collateralUsd > 0n ? (sizeInUsd * 10000n) / collateralUsd : undefined;
 
   return {
     key,
-    contractKey: `${key}:contract`,
+    contractKey: hashedPositionKey(account, marketInfo.marketTokenAddress, collateralToken.address, isLong),
     account,
     marketAddress: marketInfo.marketTokenAddress,
     collateralTokenAddress: collateralToken.address,
     sizeInUsd,
     sizeInTokens,
-    collateralAmount,
+    collateralAmount: resolvedCollateralAmount,
     pendingBorrowingFeesUsd: 0n,
     increasedAtTime: 0n,
     decreasedAtTime: 0n,
@@ -76,7 +86,7 @@ export function createMockPositionInfo({
     liquidationPrice: isLong ? (entryPrice * 105n) / 200n : (entryPrice * 295n) / 200n,
     collateralUsd,
     remainingCollateralUsd: collateralUsd,
-    remainingCollateralAmount: collateralAmount,
+    remainingCollateralAmount: resolvedCollateralAmount,
     hasLowCollateral: false,
     pnlPercentage: 0n,
     pnlAfterFees: 0n,
